Refactor query module
