+++ ocular/Cargo.toml
@@ -47,3 +47,4 @@ toml = "0.5.8"
 chrono = "0.4.0"
 uuid = { version = "0.8.2", features = ["serde", "v4"] }
 eyre = "0.6.8"
+async-trait = "0.1.56"
