+++ ocular/src/chain/client/authz.rs
@@ -1,98 +1,19 @@
 use crate::{
     account::AccountInfo,
     cosmos_modules::{
-        authz::{self, *},
+        authz::*,
         feegrant::{BasicAllowance, MsgGrantAllowance},
     },
-    error::{ChainClientError, GrpcError, TxError},
+    error::ChainClientError,
     tx::TxMetadata,
 };
 use cosmrs::{tx, AccountId};
 use prost::Message;
 use tendermint_rpc::endpoint::broadcast::tx_commit::Response;
-use tonic::transport::Channel;
 
 use super::ChainClient;
 
-pub type AuthzQueryClient = authz::query_client::QueryClient<Channel>;
-
 impl ChainClient {
-    // Authz queries
-    // TODO: Refractor code accross grpc clients by having keystore implement Sync (https://github.com/PeggyJV/ocular/pull/53#discussion_r880698565)
-    pub async fn get_authz_query_client(&mut self) -> Result<AuthzQueryClient, ChainClientError> {
-        let mut result: Result<AuthzQueryClient, ChainClientError> =
-            Err(TxError::Broadcast(String::from("Client connection never attempted.")).into());
-
-        // Get grpc address randomly each time; shuffles on failures
-        for _i in 0u8..self.connection_retry_attempts + 1 {
-            // Attempt to use last healthy (or manually set) endpoint if it exists (in config)
-            let endpoint: String = if !self.config.grpc_address.is_empty() {
-                self.config.grpc_address.clone()
-            } else {
-                // Get a random endpoint from the cache
-                match self.get_random_grpc_endpoint().await {
-                    Ok(endpt) => endpt,
-                    Err(err) => return Err(GrpcError::MissingEndpoint(err.to_string()).into()),
-                }
-            };
-
-            result = AuthzQueryClient::connect(endpoint.clone())
-                .await
-                .map_err(|e| GrpcError::Connection(e).into());
-
-            // Return if result is valid client, or increment failure in cache if being used
-            if result.is_ok() {
-                // Reset consecutive failed connections to 0
-                self.cache
-                    .as_mut()
-                    .unwrap()
-                    .grpc_endpoint_cache
-                    .add_item(endpoint.clone(), 0)?;
-
-                // Update config to last healthy grpc connection address
-                self.config.grpc_address = endpoint.clone();
-
-                break;
-            } else if result.is_err() && self.cache.is_some() {
-                // Don't bother updating config grpc address if it fails, it'll be overriden upon a successful connection
-                let _res = self
-                    .cache
-                    .as_mut()
-                    .unwrap()
-                    .grpc_endpoint_cache
-                    .increment_failed_connections(endpoint)?;
-            }
-        }
-
-        result
-    }
-
-    // Query for a specific msg grant
-    pub async fn query_authz_grant(
-        &mut self,
-        granter: &str,
-        grantee: &str,
-        msg_type_url: &str,
-    ) -> Result<QueryGrantsResponse, ChainClientError> {
-        let mut query_client = self.get_authz_query_client().await?;
-
-        let request = QueryGrantsRequest {
-            granter: granter.to_string(),
-            grantee: grantee.to_string(),
-            msg_type_url: msg_type_url.to_string(),
-            // TODO: Support pagination if use case arises
-            pagination: None,
-        };
-
-        let response = query_client
-            .grants(request)
-            .await
-            .map_err(GrpcError::Request)?
-            .into_inner();
-
-        Ok(response)
-    }
-
     // Grant Authorization
     // TODO: support other types of authorization grants other than GenericAuthorization for send messages.
     pub async fn grant_send_authorization(
@@ -200,24 +121,5 @@ impl ChainClient {
         let tx_body = tx::Body::new(vec![msg_any], &tx_metadata.memo, tx_metadata.timeout_height);
 
         self.sign_and_send_msg(granter, tx_body, tx_metadata).await
-    }
-}
-
-// Disclaimer on testing: Since the above commands inherently require chains to operate, testing is primarily deferred to integration tests in ocular/tests/single_node_chain_txs.rs
-
-#[cfg(test)]
-mod tests {
-    use crate::chain::{self, client::ChainClient};
-    use assay::assay;
-
-    #[assay]
-    async fn gets_authz_client() {
-        let mut client = ChainClient::create(chain::COSMOSHUB).unwrap();
-        client.config.grpc_address = "http://cosmoshub.strange.love:9090".to_string();
-
-        client
-            .get_authz_query_client()
-            .await
-            .expect("failed to get authz query client");
     }
 }
